requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
passlib==1.7.4
Pillow==9.3.0
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-pptx==0.6.21
pytz==2022.7
reportlab==3.6.12
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
XlsxWriter==3.0.5
